@@ -17,15 +17,15 @@
     "e2e": "ng e2e"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.803.22",
-    "@angular/cli": "8.3.27",
-    "@angular/compiler-cli": "8.2.14",
-    "@angular/language-service": "8.2.14",
+    "@angular-devkit/build-angular": "~0.901.9",
+    "@angular/cli": "9.1.9",
+    "@angular/compiler-cli": "9.1.11",
+    "@angular/language-service": "9.1.11",
     "@types/crypto-js": "3.1.43",
     "@types/jasmine": "~3.5.0",
     "@types/jasminewd2": "~2.0.8",
-    "@types/node": "~13.1.5",
-    "codelyzer": "~5.2.1",
+    "@types/node": "^12.11.1",
+    "codelyzer": "^5.1.2",
     "jasmine-core": "~3.5.0",
     "jasmine-spec-reporter": "~4.2.1",
     "karma": "~4.4.1",
@@ -35,25 +35,24 @@
     "karma-jasmine": "~3.0.3",
     "karma-jasmine-html-reporter": "1.5.1",
     "moment-es6": "1.0.0",
-    "ng-packagr": "5.7.1",
+    "ng-packagr": "^9.0.0",
     "npm-check-updates": "^4.0.1",
     "npm-install-peers": "1.2.1",
     "protractor": "~5.4.2",
     "ts-node": "~8.5.4",
-    "tsickle": "0.38.0",
     "tslint": "5.20.1",
     "typescript": "3.8.3",
-    "@angular/animations": "8.2.14",
+    "@angular/animations": "9.1.11",
     "@angular/cdk": "8.2.3",
-    "@angular/common": "8.2.14",
-    "@angular/compiler": "8.2.14",
-    "@angular/core": "8.2.14",
-    "@angular/forms": "8.2.14",
+    "@angular/common": "9.1.11",
+    "@angular/compiler": "9.1.11",
+    "@angular/core": "9.1.11",
+    "@angular/forms": "9.1.11",
     "@angular/http": "7.2.16",
     "@angular/material": "8.2.3",
-    "@angular/platform-browser": "8.2.14",
-    "@angular/platform-browser-dynamic": "8.2.14",
-    "@angular/router": "8.2.14",
+    "@angular/platform-browser": "9.1.11",
+    "@angular/platform-browser-dynamic": "9.1.11",
+    "@angular/router": "9.1.11",
     "@covalent/core": "3.0.1",
     "crypto-js": "3.1.9-1",
     "angular2-text-mask": "9.0.0",
@@ -65,7 +64,10 @@
     "ngx-device-detector": "1.3.20",
     "core-js": "2.5.7",
     "rxjs": "6.5.4",
-    "zone.js": "0.10.2",
+    "zone.js": "~0.10.2",
     "moment": "2.24.0"
+  },
+  "dependencies": {
+    "tslib": "^1.10.0"
   }
 }
